ml/requirements.txt
numpy==1.23.5
pandas==1.5.3
scikit-learn==1.2.2
joblib==1.2.0
SQLAlchemy==1.4.46
psycopg2-binary==2.9.6  # SQLAlchemy uses this as the PostgreSQL driver
